Install gccgo package instead of gccgo-go
Like gccgo-go, the gccgo package installs both the go and gccgo
compilers, but at a higher version, 1.14.6, as compared to 1..13.8 for
gccgo-go. The ggcgo package installs the go executable
x86_64-linux-gnu-go-10 in /usr/bin and creates a soft link of
/usr/bin/go-10 for it, but it doesn't also create a /usr/bin/go soft
link which appears to be a bug. Fix it by creating the link ourselves.

Signed-off-by: Shay Gordon <moebius.le.roi@gmail.com>